+++ cli/request_test.go
@@ -3,13 +3,12 @@ package cli
 import (
 	"io"
 	"net/http"
-	"os"
 	"strings"
 	"testing"
 
 	"github.com/lunjon/http/internal/client"
+	"github.com/lunjon/http/internal/config"
 	"github.com/lunjon/http/internal/logging"
-	"github.com/lunjon/http/internal/mock"
 	"github.com/spf13/cobra"
 	"github.com/stretchr/testify/require"
 )
@@ -21,46 +20,47 @@ type state struct {
 type fixture struct {
 	handler *RequestHandler
 	root    *cobra.Command
+	logs    *strings.Builder
 	infos   *strings.Builder
 	errors  *strings.Builder
 	state   *state
 }
 
-func setupRequestTest(t *testing.T) *fixture {
-	logger := logging.NewLogger()
-	logger.SetOutput(io.Discard)
+func setupRequestTest(t *testing.T, cfgs ...config.Config) *fixture {
+	logs := &strings.Builder{}
+	infos := &strings.Builder{}
+	errors := &strings.Builder{}
+
+	logger := logging.New(io.Discard)
+	logger.SetOutput(logs)
+
 	c := client.NewClient(testServer.Client(), logger, logger)
 
 	state := &state{}
+
 	failFunc := func(int) {
 		state.failCalled = true
 	}
 
-	fm := &formatterMock{}
-	sm := &signerMock{}
+	formatter := &formatterMock{}
+	signer := &signerMock{}
 
-	infos := &strings.Builder{}
-	errors := &strings.Builder{}
-
-	cfg := config{
-		fail:           false,
-		repeat:         1,
-		defaultHeaders: "x-custom: value | authorization: bearer token",
-		headerOpt:      newHeaderOption(),
-		aliasFilepath:  testAliasFilepath,
-		logs:           io.Discard,
-		infos:          infos,
-		errs:           errors,
+	cfg := config.New()
+	if len(cfgs) == 1 {
+		cfg = cfgs[0]
 	}
 
 	handler := newHandler(
 		c,
-		mock.NewManagerMock(),
-		fm,
-		sm,
+		make(map[string]string),
+		formatter,
+		signer,
 		logger,
+		cfg,
+		http.Header{},
+		infos,
+		"",
 		failFunc,
-		&cfg,
 	)
 
 	return &fixture{
@@ -81,23 +81,13 @@ func TestGet(t *testing.T) {
 }
 
 func TestGetErrorWithFail(t *testing.T) {
-	fixture := setupRequestTest(t)
-	fixture.handler.fail = true
+	fixture := setupRequestTest(t, config.New().UseFail(true))
 
 	err := fixture.handler.handleRequest("get", testServer.URL+"/error", "")
 	require.NoError(t, err)
 	require.True(t, fixture.state.failCalled)
 	require.Empty(t, fixture.infos.String())
 	require.Empty(t, fixture.errors.String())
-}
-
-func TestWithHeaders(t *testing.T) {
-	fixture := setupRequestTest(t)
-	os.Setenv("DEFAULT_HEADERS", "x-custom: value | authorization: bearer token")
-
-	err := fixture.handler.handleRequest("get", testServer.URL, "")
-	require.NoError(t, err)
-	require.NotEmpty(t, fixture.infos.String())
 }
 
 func TestPost(t *testing.T) {
@@ -114,13 +104,4 @@ func TestPost(t *testing.T) {
 		require.NotEmpty(t, fixture.infos.String())
 		require.Empty(t, fixture.errors.String())
 	}
-}
-
-func TestGetDefaultHeaders(t *testing.T) {
-	fixture := setupRequestTest(t)
-	header, err := fixture.handler.getHeaders()
-	require.NoError(t, err)
-	require.GreaterOrEqual(t, len(header), 3)
-	require.Contains(t, header, "X-Custom")
-	require.Contains(t, header, userAgentHeader)
 }