+++ tui/help.go
@@ -9,6 +9,7 @@ import (
 // keyMap defines a set of keybindings. To work for help it must satisfy
 // key.Map. It could also very easily be a map[string]key.Binding.
 type keyMap struct {
+	Enter key.Binding
 	Up    key.Binding
 	Down  key.Binding
 	Left  key.Binding
@@ -28,12 +29,16 @@ func (k keyMap) ShortHelp() []key.Binding {
 // key.Map interface.
 func (k keyMap) FullHelp() [][]key.Binding {
 	return [][]key.Binding{
-		{k.Up, k.Down, k.Left, k.Right}, // first column
-		{k.Auto, k.Help, k.Quit},        // second column
+		{k.Up, k.Down, k.Left, k.Right},   // first column
+		{k.Enter, k.Auto, k.Help, k.Quit}, // second column
 	}
 }
 
 var defaultKeyMap = keyMap{
+	Enter: key.NewBinding(
+		key.WithKeys("enter"),
+		key.WithHelp("/enter", "confirm"),
+	),
 	Up: key.NewBinding(
 		key.WithKeys("up", "k"),
 		key.WithHelp("↑/k", "move up"),